+++ src/lib/api/base.js
@@ -4,7 +4,7 @@ import { AuthenticationRoutes, FrontSiteRoutes } from '@routes/urls';
 import axios from 'axios';
 
 const apiConfig = {
-  baseURL: invenioConfig.APP.REST_ENDOINTS_BASE_URL,
+  baseURL: invenioConfig.APP.REST_ENDPOINTS_BASE_URL,
   withCredentials: true,
   xsrfCookieName: 'csrftoken',
   xsrfHeaderName: 'HTTP_X_CSRFTOKEN',