+++ src/lib/modules/Patron/PatronBulkExtendLoans/index.js
@@ -1,0 +1,16 @@
+import { connect } from 'react-redux';
+import PatronBulkExtendLoansComponent from './PatronBulkExtendLoans';
+import { bulkLoanExtension } from './actions';
+
+const mapStateToProps = (state) => ({
+  isLoading: state.bulkLoanExtend.isLoading,
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  bulkLoanExtension: (patronPid) => dispatch(bulkLoanExtension(patronPid)),
+});
+
+export const PatronBulkExtendLoans = connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(PatronBulkExtendLoansComponent);