+++ CHANGES.md
@@ -1,5 +1,23 @@
 # Changes
 
+Version 1.0.0-alpha.56 (released 2021-10-08)
+
+- add links to the search guide
+- disable the bulk extend loans if no loans to extend
+- document details page overhaul
+
+Version 1.0.0-alpha.55 (released 2021-10-08)
+
+- fix purchase order assignment to document request
+
+Version 1.0.0-alpha.54 (released 2021-09-30)
+
+- add bulk extension feature
+
+Version 1.0.0-alpha.53 (released 2021-09-28)
+
+- global: notification module integration
+
 Version 1.0.0-alpha.52 (released 2021-08-30)
 
 - document: add purchase orders to display